remove temperature_calculation_cpa.nb
Turns out that it's a pretty rough estimate that might not be too
reliable (our model for the specific heat might not work too well up to
the high temperature we care about, i.e., 1.5K).

Moreover, estimation of thermal the properties of other parts of the
system -- epoxy, salt-epoxy interface, pill-thermometer interface,
etc. -- are simply beyond our reach, rendering the results in this
notebook almost useless for determining the lowest temperature we could
get in our experiment.